% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{extract_property}
\alias{extract_property}
\title{Extract properties from an object}
\usage{
extract_property(obj, arr, propName)
}
\arguments{
\item{obj}{Object to traverse}

\item{arr}{Array of property values to look at e.g. meetings$12029, meetings$12030}

\item{propertyName}{Property to extract from each traversed object}
}
\description{
Extract properties from an object
}
\examples{
extract_property( all_meetings, meetingIds, 'name' )
extract_property( { '12345': { 'name': 'x' } }, [ 12345 ], 'name' ) > [ x ]
}
\keyword{json,}
\keyword{utils}

