% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{retrieve_meeting_id}
\alias{retrieve_meeting_id}
\title{Retrieve meeting ID}
\usage{
retrieve_meeting_id(date, animal, course)
}
\arguments{
\item{date}{date of meeting}

\item{animal}{venue type}

\item{course}{venue name}
}
\description{
This function allows you to express your love of cats.
}
\examples{
retrieveMeeting()
}
\keyword{meeting}

