% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{retrieve_meetings}
\alias{retrieve_meetings}
\title{Retrieve meeting events}
\usage{
retrieve_meetings(date, animal)
}
\arguments{
\item{meetingId}{meeting ID to use for retrival}
}
\description{
Retrieve meeting events
}
\examples{
retrieve_meetings()
}
\keyword{events}

